Carga Imagenes, Estilo Producto y Crud Categoria

+++ Pedidos.Abstracciones/ModelosParaUI/ProductoDto.cs
@@ -15,8 +15,8 @@ namespace Pedidos.Abstracciones.ModelosParaUI
 		[Required]
 		public string Nombre { get; set; }
 		public int CategoriaId { get; set; }
-		public float Precio { get; set; }
-		public float ImpuestoPorc { get; set; }
+		public decimal Precio { get; set; }
+		public decimal ImpuestoPorc { get; set; }
 		public int Stock { get; set; }
 		public bool Activo { get; set; }
 		public string ImagenUrl { get; set; }
